+++ src/bract/core/dev_init.clj
@@ -11,20 +11,34 @@
   "This namespace is solely for DEV mode initialization, typically suited for running tests from the command-line.
   To use this, simply 'require' it in the 'ns' macro of the namespace that expects DEV mode initialization:
 
+  ```clojure
   (ns myapp
     (:require bract.core.dev-init))
+  ```
 
   A more elaborate (contrived) example is below:
 
+  ```clojure
   (ns mycorp.myapp.foo-test
     (:require
-      [mycorp..myapp.foo :as f]
+      [mycorp.myapp.foo :as f]
       [clojure.test :refer [deftest is testing]]
       bract.core.dev-init))
+  ```
 
   In the snippets above the ns block ensures the initialized DEV environment by referring to `bract.core.dev-init`."
   (:require
-    [bract.core.dev :as core-dev]))
+    [clojure.java.io   :as io]
+    [bract.core.dev    :as core-dev]
+    [bract.core.keydef :as kdef]
+    [bract.core.util   :as util]))
 
 
-(core-dev/ensure-init)
+(let [context-file (kdef/ctx-context-file (core-dev/initial-context))]
+  (if (and (string? context-file)
+        (or (io/resource context-file)
+          (.exists (io/file context-file))))
+    (core-dev/ensure-init)
+    (util/err-print-banner
+      "ERROR: Context file" (pr-str context-file)
+      "not found in classpath or filesystem - ignoring DEV initialization")))